PyBob/requirements.txt
PySide6>=6.8.0
mss>=10.0.0
keyboard>=0.13.5
httpx>=0.28.0
PyYAML>=6.0.2
pynput>=1.7.7
